requirements.txt
# ArXiv API 交互
arxiv>=2.1.0

# 配置文件解析
pyyaml>=6.0

# HTTP 请求
requests>=2.31.0

# 数据处理
pandas>=2.0.0

# 定时任务调度
schedule>=1.2.0

# 日期时间处理
python-dateutil>=2.8.0

# 日志增强
colorlog>=6.7.0

# 环境变量管理（Serverless部署）
python-dotenv>=1.0.0

# 可选依赖（用于 Markdown 转 HTML，非必需）
# markdown>=3.4.0
